backend/core/requirements.txt
Flask==1.1.4
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.20
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Cors==3.0.9
requests==2.28.1
mysqlclient==2.1.1
markupsafe==2.0.1
itsdangerous<2.0.1
jinja2<3.1.0
werkzeug<2.0.1
pika==1.3.1
pytz==2022.7.1
python_dotenv